Move EDGE3D setCountedAttributes to another file
This allows us to include the unmodified EDGE3D_HELPER.hpp
in EDGE3D.cpp
