+++ src/apps/EDGE3D.cpp
@@ -120,59 +120,5 @@ void EDGE3D::tearDown(VariantID vid, size_t RAJAPERF_UNUSED_ARG(tune_idx))
   deallocData(m_sum, vid);
 }
 
-
-// Only define setCountedAttributes functions past this point
-// BEWARE: data types (Index_type, Real_ptr, etc) become wrappers past this point
-#include "common/CountingMacros.hpp"
-
-
-} // end namespace apps
-} // end namespace rajaperf
-
-// This shouldn't result in ODR violations as the argument types have changed
-#include "EDGE3D_HELPER.hpp"
-
-namespace rajaperf
-{
-namespace apps
-{
-
-void EDGE3D::setCountedAttributes()
-{
-  VariantID vid = VariantID::Base_Seq;
-  size_t tune_idx = 0;
-
-  RAJAPERF_COUNTERS_INITIALIZE();
-
-  RAJAPERF_COUNTERS_CODE_WRAPPER(
-  setUp(vid, tune_idx);
-  );
-
-  {
-    RAJAPERF_COUNTERS_CODE_WRAPPER(
-    const Index_type ibegin = m_domain->fpz;
-    const Index_type iend = m_domain->lpz+1;
-
-    EDGE3D_DATA_SETUP;
-    );
-
-    RAJAPERF_COUNTERS_REP_SCOPE()
-    {
-
-      RAJAPERF_COUNTERS_PAR_LOOP(for (Index_type i = ibegin; i < iend; ++i )) {
-        RAJAPERF_COUNTERS_LOOP_BODY(EDGE3D_BODY);
-      }
-
-    }
-
-  }
-
-  RAJAPERF_COUNTERS_CODE_WRAPPER(
-  tearDown(vid, tune_idx);
-  );
-
-  RAJAPERF_COUNTERS_FINALIZE();
-}
-
 } // end namespace apps
 } // end namespace rajaperf
